ref: Restructured order page to less components
